+++ src/main/java/com/hyerijang/dailypay/member/service/UserService.java
@@ -1,8 +1,8 @@
 package com.hyerijang.dailypay.member.service;
 
-import com.hyerijang.dailypay.member.UserRepository;
 import com.hyerijang.dailypay.member.dto.RegisterRequest;
 import com.hyerijang.dailypay.member.entity.User;
+import com.hyerijang.dailypay.member.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
